requirements.txt
# Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.23

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# LLM
anthropic==0.18.0
openai>=1.0.0

# Utilities
python-dotenv==1.0.0
aiohttp==3.9.1

# Video Processing (reserved)
# moviepy==1.0.3
